+++ java.sdk.gibra/src/main/java/gibra/scanner/KB.java
@@ -6,8 +6,12 @@ import org.apache.commons.cli.CommandLine;
 public class KB {
     private static Scanner kb = new Scanner(System.in);
 
-    public static String get(CommandLine cmd, String method) {
-        System.out.print("Enter a value for '" + method + "': ");
+    public static String get(CommandLine cmd, String option) {
+        String flag = cmd.getOptionValue(option, "");
+        if (flag != "") {
+            return flag;
+        }
+        System.out.print("Enter a value for '" + option + "': ");
         return kb.nextLine();
     }